Fix incorrect type id return for DOUBLE type in DatabaseMetaDataTest utility method.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@509261 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/functionTests/tests/jdbcapi/DatabaseMetaDataTest.java
@@ -2075,7 +2075,7 @@ public class DatabaseMetaDataTest extends BaseJDBCTestCase {
             return Types.REAL;
 
         if ("DOUBLE".equals(type) || "DOUBLE PRECISION".equals(type))
-            return Types.INTEGER;
+            return Types.DOUBLE;
         
         if ("DATE".equals(type))
             return Types.DATE;
